daemon: reuse shared env guard in systemd tests
