+++ crates/daemon/src/lib.rs
@@ -165,6 +165,9 @@ mod daemon;
 mod error;
 mod systemd;
 
+#[cfg(test)]
+mod test_env;
+
 pub use cli::{exit_code_from, run};
 pub use config::{DaemonConfig, DaemonConfigBuilder};
 pub use daemon::run_daemon;
